Move standard-responses.ts one level up
The `common` folder contained only one file, and the other things in `backend` seem to be as `common` as this one.
